tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.9.0",
    "python-dotenv>=1.1.0",
]
